Adiciona endpoint, DTOs e Feign Client para testar de forma simples

+++ src/main/java/br/com/escola/infra/aluno/AlunoConverter.java
@@ -14,7 +14,7 @@ public class AlunoConverter {
     private EncriptadorSenha encriptador;
 
     public Aluno from(AlunoEntity alunoEntity) {
-        Cpf cpf = Cpf.of(alunoEntity.getBaseCpf(), alunoEntity.getDigitosCpf());
+        Cpf cpf = Cpf.of(alunoEntity.getBaseCpf(), alunoEntity.getDigitosCpf().intValue());
 
         List<Telefone> telefones = alunoEntity.getTelefones()
                 .stream()
